sam-app/whats_new/requirements.txt
requests
beautifulsoup4
python-pptx
boto3